+++ noodles-cram/src/crai.rs
@@ -5,12 +5,14 @@ pub mod r#async;
 
 pub mod io;
 pub mod record;
-mod writer;
+
+pub use self::record::Record;
 
 #[deprecated(since = "0.74.0", note = "Use `crai::io::Reader` instead.")]
 pub use self::io::Reader;
 
-pub use self::{record::Record, writer::Writer};
+#[deprecated(since = "0.74.0", note = "Use `crai::io::Writer` instead.")]
+pub use self::io::Writer;
 
 #[cfg(feature = "async")]
 pub use self::r#async::{Reader as AsyncReader, Writer as AsyncWriter};