+++ noodles-cram/src/crai/io.rs
@@ -1,5 +1,6 @@
 //! CRAM index I/O.
 
 mod reader;
+mod writer;
 
-pub use self::reader::Reader;
+pub use self::{reader::Reader, writer::Writer};